Retrieve Part I from first 100 companies in S&P 500 retrieve_part1.py was executed over the last 5 filings for the first 100 companies in the S&P 500. This was done to greatly reduce the file size for each data input. Currently there are a handful of companies that raised errors, so there should be another preprocessing step to eliminate repeated filings, or handle fewer than 5 filings.
The data is stored in a directory named after the compay's ticker
symbol, and each file is named with the ticker, accompanied by the
date retrieved that the data covers up to. e.g. Ticker_20201231.txt
indicates the 10-K was for the year 2020 from January to the end of
December.